requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
pandas
pyarrow
loguru
sqlalchemy
pendulum